修复数据库初始化问题、删除过时的数据库格式方法、修复空停用词错误。
